Working on implementing Dask cluster for distributed computations
I have an almost working proof of concept for implementing a Dask distributed cluster when performing the `fitness_func`. This is similar to the parallel processing recently implemented, however this allows the user to distribute that beyond their local machine using their Dask cluster of machines/processes.

I'm currently stuck trying to work on debugging a `distributed` bug that I can't make sense of why it's happening. I'll keep working on it.